+++ x/ibc/23-commitment/types.go
@@ -5,6 +5,6 @@ type Root interface{}
 
 // XXX: need to separate membership and non membership proof types
 type Proof interface {
-	Key() []byte
+	GetKey() []byte
 	Verify(Root, []byte) error
 }
